requirements.txt
torch==1.13.1
torchvision==0.14.1
Pillow==9.5.0
fastapi
uvicorn
requests